remove and rename config.ts from utils to src
